+++ lib/domain/calendar/calendar_page.dart
@@ -6,7 +6,7 @@ import 'package:pilll/domain/calendar/calendar_page_index_state_notifier.dart';
 import 'package:pilll/domain/calendar/calendar_page_state.codegen.dart';
 import 'package:pilll/domain/calendar/components/title/calendar_page_title.dart';
 import 'package:pilll/domain/calendar/components/month_calendar/month_calendar.dart';
-import 'package:pilll/components/organisms/calendar/weekly/weekly_calendar.dart';
+import 'package:pilll/components/organisms/calendar/week/week_calendar.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/domain/calendar/components/pill_sheet_modified_history/pill_sheet_modified_history_card.dart';
 import 'package:pilll/domain/calendar/calendar_page_state_notifier.dart';
@@ -111,7 +111,7 @@ class CalendarPageBody extends StatelessWidget {
                             calendarNextPillSheetBandModels:
                                 state.calendarNextPillSheetBandModels,
                             horizontalPadding: 0,
-                            onTap: (weeklyCalendarState, date) {
+                            onTap: (weekCalendarState, date) {
                               analytics.logEvent(
                                   name: "did_select_day_tile_on_calendar_card");
                               transitionToPostDiary(
