+++ lib/domain/menstruation/components/calendar/menstruation_single_line_state.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pilll/components/organisms/calendar/band/calendar_band_model.dart';
 import 'package:pilll/domain/calendar/date_range.dart';
-import 'package:pilll/components/organisms/calendar/weekly/weekly_calendar_state.dart';
+import 'package:pilll/components/organisms/calendar/week/week_calendar_state.dart';
 import 'package:pilll/entity/diary.codegen.dart';
 
 class MenstruationSinglelineWeeklyCalendarState extends WeekCalendarState {
